+++ examples/arm_hand_xarm_leap.py
@@ -91,7 +91,7 @@ if __name__ == "__main__":
     ]
 
     # IK settings.
-    solver = "quadprog"
+    solver = "daqp"
     model = configuration.model
     data = configuration.data