+++ examples/arm_ur5e.py
@@ -50,7 +50,7 @@ if __name__ == "__main__":
     mid = model.body("target").mocapid[0]
     model = configuration.model
     data = configuration.data
-    solver = "quadprog"
+    solver = "daqp"
 
     with mujoco.viewer.launch_passive(
         model=model, data=data, show_left_ui=False, show_right_ui=False