+++ examples/humanoid_h1.py
@@ -59,7 +59,7 @@ if __name__ == "__main__":
 
     model = configuration.model
     data = configuration.data
-    solver = "quadprog"
+    solver = "daqp"
 
     with mujoco.viewer.launch_passive(
         model=model, data=data, show_left_ui=False, show_right_ui=False