+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Added support for Python 3.8.
+- Added equality operators for SE3 and SO3.
 
 ## [0.0.6] - 2025-03-15